+++ src/main/java/com/tarikh/interview/starling/domain/models/GoalContainer.java
@@ -8,7 +8,7 @@ import lombok.extern.slf4j.Slf4j;
 @Builder
 @Data
 public class GoalContainer {
-    private final double amountToAdd;
+    private final int amountToAdd;
     private final String accUId;
     private final String nameOfGoal;
     private final String goalId;
